x/fractal/types/message_settle_exchange_test.go
package types

import (
	"testing"

	"fractal/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgSettleExchange_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSettleExchange
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSettleExchange{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSettleExchange{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
